Tests/LibWeb/Text/expected/ResizeObserver/unobserve.txt

Size changed: 200px x 200px
